+++ model/src/main/java/net/artux/pda/model/quest/TransferModel.kt
@@ -1,3 +1,3 @@
 package net.artux.pda.model.quest
 
-data class TransferModel(val stageId: Int, val text: String)
+data class TransferModel(val stageId: Long, val text: String)
